fixed pointer problem om 64 bits systems
git-svn-id: https://ilk.uvt.nl/svn/trunk/sources/Timbl6@3405 12f355fe-0486-481a-ad91-c297ab22b4e3

+++ demos/sockettestServer.cxx
@@ -49,8 +49,8 @@ void *do_child( void *arg ){
   // process the test material
   //
   char line[256];
-  sprintf( line, "Server: Thread %u, on Socket %d", 
-	   (unsigned int)pthread_self(),
+  sprintf( line, "Server: Thread %lu, on Socket %d", 
+	   (uintptr_t)pthread_self(),
 	   mysock->getSockId() );
   cerr << line << ", started" << endl;
   string buf;
@@ -85,8 +85,8 @@ void *do_to_child( void *arg ){
   // process the test material
   //
   char line[256];
-  sprintf( line, "ToServer: Thread %u, on Socket %d", 
-	   (unsigned int)pthread_self(),
+  sprintf( line, "ToServer: Thread %lu, on Socket %d", 
+	   (uintptr_t)pthread_self(),
 	   mysock->getSockId() );
   cerr << line << ", started" << endl;
   string buf;
